+++ CHANGELOG.md
@@ -7,15 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.49.0] - 2025-11-12
+
 ### Added
 
 - Add Alertmanager routes unit tests
 - Add Alertmanager routes integration tests
   test scenarios: heartbeat alerts, pipeline stable and testing alerts, Slack notifications, and ticket creation.
 
+### Changed
+
+- Refactored Grafana package to use domain organization objects and moved status updates to controller layer.
+- Refactor webhook test suite architecture to use shared testutil package and individual test suites per API version
+
 ### Removed
 
 - Remove prometheus-agent support as we now fully run alloys.
+- Remove all opsgenie heartbeats.
 
 ## [0.48.1] - 2025-11-05
 
@@ -754,7 +762,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initialize project and create heartbeat for the installation.
 
-[Unreleased]: https://github.com/giantswarm/observability-operator/compare/v0.48.1...HEAD
+[Unreleased]: https://github.com/giantswarm/observability-operator/compare/v0.49.0...HEAD
+[0.49.0]: https://github.com/giantswarm/observability-operator/compare/v0.48.1...v0.49.0
 [0.48.1]: https://github.com/giantswarm/observability-operator/compare/v0.48.0...v0.48.1
 [0.48.0]: https://github.com/giantswarm/observability-operator/compare/v0.47.0...v0.48.0
 [0.47.0]: https://github.com/giantswarm/observability-operator/compare/v0.46.2...v0.47.0